+++ client.go
@@ -196,7 +196,7 @@ func (c *Client) AuthAPI() http.Client {
 	return client
 }
 
-// Dedicated databases APIs are actually available in preview only.
+// Preview returns the dedicated databases APIs client, available in preview.
 func (c *Client) Preview() PreviewClient {
 	return NewPreviewClient(c)
 }