+++ databases_preview.go
@@ -7,7 +7,7 @@ import (
 	"github.com/Scalingo/go-utils/errors/v2"
 )
 
-const databasesPreviewResource = "databases"
+const databasesResource = "databases"
 
 var ErrDatabaseNotFound = stderrors.New("database not found")
 
@@ -54,7 +54,7 @@ func NewPreviewClient(parent *Client) PreviewClient {
 
 func (c *PreviewClient) DatabaseCreate(ctx context.Context, params DatabaseCreateParams) (DedicatedDatabase, error) {
 	var res DedicatedDatabase
-	err := c.parent.ScalingoAPI().ResourceAdd(ctx, databasesPreviewResource, params, &res)
+	err := c.parent.ScalingoAPI().ResourceAdd(ctx, databasesResource, params, &res)
 	if err != nil {
 		return res, errors.Wrap(ctx, err, "create dedicated database")
 	}
@@ -64,20 +64,20 @@ func (c *PreviewClient) DatabaseCreate(ctx context.Context, params DatabaseCreat
 func (c *PreviewClient) DatabasesList(ctx context.Context) ([]DedicatedDatabase, error) {
 	var res []DedicatedDatabase
 
-	err := c.parent.ScalingoAPI().ResourceList(ctx, databasesPreviewResource, nil, &res)
+	err := c.parent.ScalingoAPI().ResourceList(ctx, databasesResource, nil, &res)
 	if err != nil {
 		return res, errors.Wrap(ctx, err, "list dedicated databases")
 	}
 	return res, nil
 }
 
-// Currently, appID is the dedicated database identifier.
+// DatabaseShow currently uses appID as the dedicated database identifier.
 func (c *PreviewClient) DatabaseShow(ctx context.Context, appID string) (DedicatedDatabaseShowResponse, error) {
 	var res DedicatedDatabaseShowResponse
 
 	dedicatedDatabase, err := c.searchDatabase(ctx, appID)
 	if err != nil {
-		return res, errors.Wrap(ctx, err, "show dedicated database")
+		return res, errors.Wrap(ctx, err, "search dedicated database")
 	}
 
 	database, err := c.parent.DatabaseShow(ctx, dedicatedDatabase.App.ID, dedicatedDatabase.Addon.ID)
@@ -92,11 +92,11 @@ func (c *PreviewClient) DatabaseShow(ctx context.Context, appID string) (Dedicat
 	return res, nil
 }
 
-// Currently, appID is the dedicated database identifier.
+// DatabaseDestroy currently uses appID as the dedicated database identifier.
 func (c *PreviewClient) DatabaseDestroy(ctx context.Context, appID string) error {
 	dedicatedDatabase, err := c.searchDatabase(ctx, appID)
 	if err != nil {
-		return errors.Wrap(ctx, err, "destroy dedicated database")
+		return errors.Wrap(ctx, err, "search dedicated database")
 	}
 
 	appName := dedicatedDatabase.App.Name
@@ -107,7 +107,7 @@ func (c *PreviewClient) DatabaseDestroy(ctx context.Context, appID string) error
 	return nil
 }
 
-// Linear database search based on DatabasesList method.
+// searchDatabase performs a linear search through DatabasesList method result.
 func (c *PreviewClient) searchDatabase(ctx context.Context, appID string) (DedicatedDatabase, error) {
 	var res DedicatedDatabase